+++ funnel/views/__init__.py
@@ -16,6 +16,7 @@ from . import (
     index,
     label,
     login,
+    login_session,
     membership,
     mixins,
     organization,
