+++ funnel/views/auth_oauth.py
@@ -18,7 +18,7 @@ from ..models import (
 from ..registry import resource_registry
 from ..utils import abort_null, make_redirect_url
 from .auth_resource import get_userinfo
-from .helpers import requires_client_login, requires_login_no_message
+from .login_session import requires_client_login, requires_login_no_message
 
 
 class ScopeException(Exception):
