+++ funnel/views/contact.py
@@ -15,7 +15,8 @@ from coaster.views import ClassView, render_with, requestargs, route
 from .. import app, funnelapp
 from ..models import ContactExchange, Participant, Project, db
 from ..utils import abort_null, format_twitter_handle
-from .helpers import app_url_for, requires_login
+from .helpers import app_url_for
+from .login_session import requires_login
 
 
 def contact_details(participant):
