+++ funnel/views/event.py
@@ -20,8 +20,8 @@ from ..models import (
     db,
 )
 from .decorators import legacy_redirect
-from .helpers import requires_login
 from .jobs import import_tickets
+from .login_session import requires_login
 from .mixins import EventViewMixin, ProjectViewMixin
