+++ funnel/views/organization.py
@@ -16,7 +16,7 @@ from .. import app
 from ..forms import OrganizationForm, TeamForm
 from ..models import Organization, Team, db
 from ..signals import org_data_changed, team_data_changed
-from .helpers import requires_login
+from .login_session import requires_login
 
 # --- Routes: Organizations ---------------------------------------------------
