+++ funnel/views/profile.py
@@ -19,7 +19,7 @@ from .. import app, funnelapp
 from ..forms import ProfileBannerForm, ProfileForm, ProfileLogoForm, SavedProjectForm
 from ..models import Profile, Project, db
 from .decorators import legacy_redirect
-from .helpers import requires_login
+from .login_session import requires_login
 from .mixins import ProfileViewMixin
